test(datetime): migrate time-label
