+++ core/src/components/datetime/test/time-label/datetime.e2e.ts
@@ -1,0 +1,31 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+configs({ modes: ['ios'], directions: ['ltr'] }).forEach(({ title, config }) => {
+  test.describe(title('datetime: time label'), () => {
+    test('should render default time label', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-datetime></ion-datetime>
+      `,
+        config
+      );
+      await page.waitForSelector('.datetime-ready');
+
+      const timeLabel = page.locator('ion-datetime .time-header');
+      await expect(timeLabel).toHaveText('Time');
+    });
+    test('should not render a custom time label', async ({ page }) => {
+      await page.setContent(
+        `
+        <ion-datetime show-default-time-label="false"></ion-datetime>
+      `,
+        config
+      );
+      await page.waitForSelector('.datetime-ready');
+
+      const timeLabel = page.locator('ion-datetime .time-header');
+      await expect(timeLabel).toHaveText('');
+    });
+  });
+});
